renaming dependencies
