+++ internal/wire/relationship/relationship_test.go
@@ -8,8 +8,8 @@ import (
 	"github.com/ardanlabs/kit/db"
 	"github.com/ardanlabs/kit/db/mongo"
 	"github.com/ardanlabs/kit/tests"
-	"github.com/coralproject/xenia/internal/shelf/relationship"
-	"github.com/coralproject/xenia/internal/shelf/relationship/relationshipfix"
+	"github.com/coralproject/xenia/internal/wire/relationship"
+	"github.com/coralproject/xenia/internal/wire/relationship/relationshipfix"
 )
 
 func init() {
